refactor: 체스 판 초기화 역할 전략 패턴으로 분리
