+++ src/main/java/chess/Board.java
@@ -3,83 +3,26 @@ package chess;
 import static chess.PieceColor.*;
 import static chess.Rank.*;
 
-import java.util.HashMap;
-import java.util.List;
-import java.util.ListIterator;
 import java.util.Map;
-import java.util.function.Function;
 
-import chess.piece.Bishop;
 import chess.piece.EmptyPiece;
 import chess.piece.King;
 import chess.piece.Knight;
 import chess.piece.Pawn;
 import chess.piece.Piece;
-import chess.piece.Queen;
-import chess.piece.Rook;
 import chess.turndecider.TurnDecider;
 
 public class Board {
 
     static final String SOURCE_POSITION_SHOULD_HAVE_PIECE_MESSAGE = "[ERROR] 출발 위치에는 말이 있어야 합니다.";
     private static final EmptyPiece EMPTY_PIECE = new EmptyPiece(EMPTY);
-    private static final Function<PieceColor, List<Piece>> createPieceFunction = (PieceColor pieceColor) -> List.of(
-        new Rook(pieceColor), new Knight(pieceColor), new Bishop(pieceColor),
-        new Queen(pieceColor), new King(pieceColor), new Bishop(pieceColor),
-        new Knight(pieceColor), new Rook(pieceColor));
 
     private final Map<Position, Piece> values;
     private final TurnDecider turnDecider;
 
-    public Board(TurnDecider turnDecider) {
-        this.values = initBoard();
+    public Board(TurnDecider turnDecider, BoardInitializer initializer) {
+        this.values = initializer.apply();
         this.turnDecider = turnDecider;
-    }
-
-    private Map<Position, Piece> initBoard() {
-        Map<Position, Piece> result = new HashMap<>();
-
-        putAllEmptyPieces(result);
-        putBlackPieces(result);
-        putWhitePieces(result);
-
-        return result;
-    }
-
-    private void putAllEmptyPieces(Map<Position, Piece> result) {
-        for (Rank rank : Rank.reverseValues()) {
-            putEmptyPiecesInOneRank(result, rank);
-        }
-    }
-
-    private void putEmptyPiecesInOneRank(Map<Position, Piece> result, Rank rank) {
-        for (File file : File.values()) {
-            result.put(new Position(rank, file), EMPTY_PIECE);
-        }
-    }
-
-    private void putBlackPieces(Map<Position, Piece> result) {
-        putPawns(result, BLACK, SEVEN);
-        putRemainPiecesExceptPawn(result, BLACK, EIGHT);
-    }
-
-    private void putWhitePieces(Map<Position, Piece> result) {
-        putPawns(result, WHITE, TWO);
-        putRemainPiecesExceptPawn(result, WHITE, ONE);
-    }
-
-    private void putRemainPiecesExceptPawn(Map<Position, Piece> result, PieceColor color, Rank rank) {
-        ListIterator<Piece> piecesIterator = createPieceFunction.apply(color).listIterator();
-
-        for (File file : File.values()) {
-            result.put(new Position(rank, file), piecesIterator.next());
-        }
-    }
-
-    private void putPawns(Map<Position, Piece> result, PieceColor color, Rank rank) {
-        for (File file : File.values()) {
-            result.put(new Position(rank, file), new Pawn(color));
-        }
     }
 
     public boolean move(Position source, Position target) {
