+++ src/main/java/chess/BoardInitializer.java
@@ -1,0 +1,9 @@
+package chess;
+
+import java.util.Map;
+
+import chess.piece.Piece;
+
+public interface BoardInitializer {
+    Map<Position, Piece> apply();
+}
